Добавит класс `board-controller`, опишет интерфейсБ перенесёт рендеринг доски и карточек из `main.js` в `controllers/board.js`
