+++ src/components/task-edit.js
@@ -1,6 +1,6 @@
 import AbstractComponent from "./abstract-component.js";
 import {COLORS, DAYS, MONTH_NAMES} from "../const.js";
-import {formatTime} from "../utils/utils.js";
+import {formatTime} from "../utils/common.js";
 
 
 const createColorsMarkup = (colors, currentColor) => {
@@ -140,5 +140,12 @@ export default class TaskEdit extends AbstractComponent {
 
   getTemplate() {
     return createTaskEditTemplate(this._task);
+  }
+
+
+  // // Добавит в компонент методы для подписки на события
+  setSubmitHandler(handler) {
+    this.getElement().querySelector(`form`)
+      .addEventListener(`submit`, handler);
   }
 }
